Разбил проект на модули, убрал необязательные файлы, поигрался с логированием
