+++ client/src/main/java/ru/gb/lefandoc/cloudstorage/client/CloudApplication.java
