+++ client/src/main/resources/log4j.properties
@@ -1,0 +1,25 @@
+# Root logger option
+log4j.rootLogger=DEBUG, StdoutDebug, StdoutInfo, StdoutWEF
+
+# Console log DEBUG level
+log4j.appender.StdoutDebug=org.apache.log4j.ConsoleAppender
+log4j.appender.StdoutDebug.Target=System.out
+log4j.appender.StdoutDebug.layout=org.apache.log4j.PatternLayout
+log4j.appender.StdoutDebug.layout.ConversionPattern=%d{HH:mm:ss dd.MM.yyyy} \u001b[1;34m[%-5p]\u001b[m %c{1}:%M:%L - %m%n
+log4j.appender.StdoutDebug.filter.debug=org.apache.log4j.varia.LevelRangeFilter
+log4j.appender.StdoutDebug.filter.debug.levelMax=DEBUG
+# Console log INFO level
+log4j.appender.StdoutInfo=org.apache.log4j.ConsoleAppender
+log4j.appender.StdoutInfp.Target=System.out
+log4j.appender.StdoutInfo.layout=org.apache.log4j.PatternLayout
+log4j.appender.StdoutInfo.layout.conversionPattern=%d{HH:mm:ss dd.MM.yyyy} \u001b[1;32m[%-5p]\u001b[m %11c{1}:%M:%L - %m%n
+log4j.appender.StdoutInfo.filter.info=org.apache.log4j.varia.LevelRangeFilter
+log4j.appender.StdoutInfo.filter.info.levelMin=INFO
+log4j.appender.StdoutInfo.filter.info.levelMax=INFO
+# Console log WARN+ level
+log4j.appender.StdoutWEF=org.apache.log4j.ConsoleAppender
+log4j.appender.StdoutWEF.Target=System.out
+log4j.appender.StdoutWEF.layout=org.apache.log4j.PatternLayout
+log4j.appender.StdoutWEF.layout.conversionPattern=%d{HH:mm:ss dd.MM.yyyy} \u001b[1;33m[%-5p]\u001b[m %11c{1}:%M:%L - %m%n
+log4j.appender.StdoutWEF.filter.wef=org.apache.log4j.varia.LevelRangeFilter
+log4j.appender.StdoutWEF.filter.wef.levelMin=WARN
