+++ client/src/main/resources/ru/gb/lefandoc/cloudstorage/client/cloud-storage.fxml
