+++ commons/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module commons {
+    requires org.slf4j;
+    requires lombok;
+    exports ru.gb.lefandoc.cloudstorage.commons.model;
+    exports ru.gb.lefandoc.cloudstorage.commons.service;
+}
