+++ commons/src/main/java/ru/gb/lefandoc/cloudstorage/commons/model/CloudMessage.java
@@ -1,0 +1,7 @@
+package ru.gb.lefandoc.cloudstorage.commons.model;
+
+import java.io.Serializable;
+
+public interface CloudMessage extends Serializable {
+    MessageType getMessageType();
+}
