+++ commons/src/main/java/ru/gb/lefandoc/cloudstorage/commons/model/FileMessage.java
@@ -1,0 +1,26 @@
+package ru.gb.lefandoc.cloudstorage.commons.model;
+
+import lombok.Data;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+@Data
+public class FileMessage implements CloudMessage {
+
+    private final byte[] bytes;
+    private final long size;
+    private final String name;
+
+    public FileMessage(Path path) throws IOException {
+        bytes = Files.readAllBytes(path);
+        size = bytes.length;
+        name = path.getFileName().toString();
+    }
+
+    @Override
+    public MessageType getMessageType() {
+        return MessageType.FILE;
+    }
+}
