+++ commons/src/main/java/ru/gb/lefandoc/cloudstorage/commons/model/MessageType.java
@@ -1,0 +1,7 @@
+package ru.gb.lefandoc.cloudstorage.commons.model;
+
+public enum MessageType {
+    FILE,
+    LIST,
+    FILE_REQUEST
+}
