+++ server/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module server {
+    requires commons;
+    requires io.netty.transport;
+    requires io.netty.codec;
+    requires org.slf4j;
+    requires lombok;
+}
