--- src/main/java/ru/gb/lefandoc/cloudstorage/client/package-info.java
@@ -1,1 +1,0 @@
-package ru.gb.lefandoc.cloudstorage.client;
