--- src/main/java/ru/gb/lefandoc/cloudstorage/server/model/CloudMessage.java
@@ -1,7 +1,0 @@
-package ru.gb.lefandoc.cloudstorage.server.model;
-
-import java.io.Serializable;
-
-public interface CloudMessage extends Serializable {
-    MessageType getMessageType();
-}
