--- src/main/java/ru/gb/lefandoc/cloudstorage/server/model/FileRequest.java
@@ -1,18 +1,0 @@
-package ru.gb.lefandoc.cloudstorage.server.model;
-
-import lombok.Data;
-
-@Data
-public class FileRequest implements CloudMessage {
-
-    private final String name;
-
-    public FileRequest(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public MessageType getMessageType() {
-        return MessageType.FILE_REQUEST;
-    }
-}
