--- src/main/java/ru/gb/lefandoc/cloudstorage/server/model/MessageType.java
@@ -1,7 +1,0 @@
-package ru.gb.lefandoc.cloudstorage.server.model;
-
-public enum MessageType {
-    FILE,
-    LIST,
-    FILE_REQUEST
-}
