--- src/main/java/ru/gb/lefandoc/cloudstorage/server/package-info.java
@@ -1,1 +1,0 @@
-package ru.gb.lefandoc.cloudstorage.server;
