--- src/main/java/ru/gb/lefandoc/cloudstorage/server/service/UserNameService.java
@@ -1,12 +1,0 @@
-package ru.gb.lefandoc.cloudstorage.server.service;
-
-public class UserNameService {
-
-    private static int userId = 0;
-
-    public String getUserName() {
-        userId++;
-        return "user" + userId;
-    }
-
-}
